refactor(web): Restrict modules, remove common
