+++ ui/web/src/modules/AccountInfo/index.ts
@@ -1,3 +1,4 @@
 import './AccountInfoPanel';
 import './AccountList';
 import './AccountPerformancePanel';
+export { useAccountInfo } from './model';
