+++ ui/web/src/modules/AccountInfo/model.ts
@@ -1,6 +1,16 @@
-import { BehaviorSubject, defer, shareReplay, switchMap } from 'rxjs';
-import { terminal$ } from '../../common/create-connection';
 import { AccountPerformanceUnit, IAccountPerformance } from '@yuants/kernel';
+import { IAccountInfo } from '@yuants/protocol';
+import { BehaviorSubject, Observable, defer, shareReplay, switchMap } from 'rxjs';
+import { terminal$ } from '../Terminals';
+
+export const useAccountInfo = (() => {
+  const hub: Record<string, Observable<IAccountInfo>> = {};
+  return (account_id: string) =>
+    (hub[account_id] ??= defer(() => terminal$).pipe(
+      switchMap((terminal) => terminal.useAccountInfo(account_id)),
+      shareReplay(1),
+    ));
+})();
 
 export const accountIds$ = defer(() => terminal$).pipe(
   switchMap((terminal) => terminal.accountIds$),
