+++ ui/web/src/modules/DesktopLayout/DesktopLayout.tsx
@@ -4,8 +4,8 @@ import { Actions, Layout as FlexLayout, TabNode } from 'flexlayout-react';
 import { useObservableState } from 'observable-hooks';
 import React, { useContext, useEffect, useState } from 'react';
 import { Trans, useTranslation } from 'react-i18next';
-import { isDarkMode$ } from '../../common/Darkmode';
-import { ErrorBoundary } from '../../common/ErrorBoundary';
+import { isDarkMode$ } from '../Workbench/darkmode';
+import { ErrorBoundary } from '../Pages/ErrorBoundary';
 import { CommandCenter, registerCommand } from '../CommandCenter';
 import { NetworkStatusWidget } from '../Terminals/NetworkStatusWidget';
 import { Login } from '../User/Login';
