+++ ui/web/src/modules/DesktopLayout/layout-model.ts
@@ -1,7 +1,7 @@
 import * as FlexLayout from 'flexlayout-react';
 import hotkeys from 'hotkeys-js';
 import { BehaviorSubject, bufferCount, combineLatest, first, map, Subject } from 'rxjs';
-import { createPersistBehaviorSubject } from '../../common/utils';
+import { createPersistBehaviorSubject } from '../FileSystem/createPersistBehaviorSubject';
 import { registerCommand } from '../CommandCenter';
 
 const initialJson = (): FlexLayout.IJsonModel => ({
