+++ ui/web/src/modules/GeneralSpecificRelations/GeneralSpecificRelationList.tsx
@@ -4,9 +4,9 @@ import { IDataRecord } from '@yuants/protocol';
 import { useObservable, useObservableState } from 'observable-hooks';
 import { useState } from 'react';
 import { combineLatest, first, mergeMap, tap, toArray } from 'rxjs';
-import { terminal$ } from '../../common/create-connection';
 import Form from '../Form';
 import { registerPage } from '../Pages';
+import { terminal$ } from '../Terminals';
 
 // TODO: Import
 interface IGeneralSpecificRelation {
