+++ ui/web/src/modules/Pages/ErrorBoundary.tsx
