+++ ui/web/src/modules/Products/index.ts
@@ -1,1 +1,2 @@
 import './ProductList';
+export { useProducts } from './model';
