+++ ui/web/src/modules/SupaBase/index.ts
@@ -1,0 +1,1 @@
+export { authState$, supabase } from './supabase';
